project/requirements.txt
celery==4.2.1
falcon==1.4.1
falcon-multipart
flower==0.9.2
gunicorn==19.9.0
redis==3.2.0
tornado>=4.2.0,<6.0.0
pillow==2.8.1